refactor(steps): move step layout and header into ui components
