+++ src/features/sell/steps/CardInformationStep.tsx
@@ -2,8 +2,8 @@ import React, { useMemo } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { StepCardInformationFormData, StepCardInformationSchema } from '../schemas/sellFormSchema';
-import StepLayout from './shared/StepLayout';
-import StepHeader from './shared/StepHeader';
+import StepLayout from '../../../components/ui/multistepsform/StepLayout';
+import StepHeader from '../../../components/ui/multistepsform/StepHeader';
 import FormSelectInput from '../../../components/ui/inputs/FormSelectInput';
 import { ActivityIndicator, View } from 'react-native';
 import { useTheme } from '../../../theme/useTheme';
