+++ src/features/sell/steps/PriceStep.tsx
@@ -3,8 +3,8 @@ import { useForm } from 'react-hook-form';
 import { StepPriceFormData, stepPriceSchema } from '../schemas/sellFormSchema';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { FormTextInput } from '../../../components/ui';
-import StepHeader from './shared/StepHeader';
-import StepLayout from './shared/StepLayout';
+import StepLayout from '../../../components/ui/multistepsform/StepLayout';
+import StepHeader from '../../../components/ui/multistepsform/StepHeader';
 import PriceEstimation from '../components/PriceEstimation';
 import { useSellForm } from '../context/SellFormContext';
 import { EuroIcon } from 'lucide-react-native';
